+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+### v2.5.0 (2019-07-19)
+
+* Added `nr.interface.Attribute(default)` parameter and `default` member
+* Added `nr.interface.Attribute(static)` parameter and `static` member
+* Added `nr.interface.Method(static)` parameter and `static` member
+* Added `nr.interface.Attribute.make_default()` parameter and `static` member
+* Changed `nr.interface.staticattr()`, now returns an `Attribute` instance
+  with `static=True`
+* Changed `nr.interface.Method.is_candidate()` to accept `staticmethod`
+  and `classmethod` instances
+* Changed `nr.interface.Method.wrap_candidate()` to accept `staticmethod`
+  and `classmethod` instances
+* Removed `nr.interface.StaticAttribute`
+
 ### v2.4.0 (2019-06-25)
 
 * Add `stream.batch()`